Change IClientTransport.Port type to ushort from int

+++ src/ScaleNet.Client/LowLevel/Transport/IClientTransport.cs
@@ -6,7 +6,7 @@ namespace ScaleNet.Client.LowLevel.Transport
     public interface IClientTransport : IDisposable
     {
         public string Address { get; }
-        public int Port { get; }
+        public ushort Port { get; }
     
         public bool ConnectClient();
         public bool ReconnectClient();
